seperated account component
