+++ frontend/src/components/navbar/RebelHubNavBar.module.css
@@ -73,27 +73,6 @@
 	width: 40px;
 	height: 40px;
 }
-
-.accountButton{
-	background: transparent;
-	display: flex;
-	justify-content: center;
-	width: 150px;
-	height: 50px;
-	border: none;
-}
-.accountButton:hover{
-	cursor: pointer;
-}
-.avatarIcon{
-	width: 30%;
-	height: 90%;
-}
-
-.accountText{
-	padding-top: 10px;
-}
-
 .extraSpace{
 	padding-top: 100px;
 }
